DB cleanup fixes. Fixed canceling notification.
Removing publications now when the author is removed. Added a clean-up on upgrade to remove all the stale publications that were left when removing authors. Selected authors are now removed in a transaction to speed things up.
Initial update on app fresh start has been delayed by 10 minutes to avoid issues on backup restore.
Now authors and publications are restored in a transaction to speed things up.