Crouton fixes. Fixed random blink of images due to network delays
Crouton in PublicationInfoFragment now shows with appropriate top margin accounting for action bar.
By default we assume there are no images, and thus image view pager is not being shown. When the first step of image retrieval completes and we know for sure that there are images to show - photo view is made visible and height is recomputed.